test: move sandbox integration tests into separate directory. (#1168)
