+++ crates/contract/tests/inprocess/mod.rs
@@ -1,0 +1,1 @@
+pub mod expired_attestation;
