+++ crates/contract/tests/sandbox/integration_tee_cleanup_after_resharing.rs
@@ -1,10 +1,9 @@
-pub mod common;
 use anyhow::Result;
 use dtos_contract::{Attestation, MockAttestation};
 use near_workspaces::{Account, Contract};
 use serde_json::json;
 
-use common::{
+use crate::sandbox::common::{
     assert_running_return_participants, check_call_success, check_call_success_all_receipts,
     gen_accounts, get_tee_accounts, init_env_secp256k1, submit_participant_info,
     submit_tee_attestations,
