+++ crates/contract/tests/sandbox/mod.rs
@@ -1,0 +1,9 @@
+pub mod back_compatibility;
+pub mod ckd;
+pub mod integration_tee_cleanup_after_resharing;
+pub mod sign;
+pub mod tee;
+pub mod updates;
+pub mod vote;
+
+pub mod common;
