+++ crates/contract/tests/sandbox/sign.rs
@@ -1,5 +1,4 @@
-pub mod common;
-use common::{
+use crate::sandbox::common::{
     candidates, create_response, init, init_env_ed25519, init_env_secp256k1, sign_and_validate,
 };
 use mpc_contract::{
