+++ crates/contract/tests/sandbox/tee.rs
@@ -1,12 +1,9 @@
-pub mod common;
-
-use crate::common::gen_accounts;
+use crate::sandbox::common::{
+    assert_running_return_participants, check_call_success, gen_accounts, get_tee_accounts,
+    init_env_ed25519, init_env_secp256k1, submit_participant_info, submit_tee_attestations,
+};
 use anyhow::Result;
 use assert_matches::assert_matches;
-use common::{
-    assert_running_return_participants, check_call_success, get_tee_accounts, init_env_ed25519,
-    init_env_secp256k1, submit_participant_info, submit_tee_attestations,
-};
 use dtos_contract::{Attestation, MockAttestation};
 use mpc_contract::{errors::InvalidState, state::ProtocolContractState};
 use mpc_primitives::hash::MpcDockerImageHash;
