+++ crates/contract/tests/sandbox/vote.rs
@@ -1,7 +1,5 @@
-pub mod common;
-
+use crate::sandbox::common::{check_call_success, gen_accounts, init_env_secp256k1};
 use assert_matches::assert_matches;
-use common::{check_call_success, gen_accounts, init_env_secp256k1};
 use mpc_contract::{
     primitives::thresholds::{Threshold, ThresholdParameters},
     state::{running::RunningContractState, ProtocolContractState},
