+++ crates/contract/tests/test.rs
@@ -1,0 +1,2 @@
+pub mod inprocess;
+pub mod sandbox;
